pyproject.toml
[tool.poetry]
name = "tempracatao"
version = "0.0.0"
description = "Project Description"
authors = ["O Boticario <you@example.com>"]

[tool.poetry.dependencies]
dj-database-url = "^0.5.0"
django = "^4.0"
django-cors-headers = "^3.3.0"
django-filter = "^21.1"
django-localflavor = "^3.0"
django-structlog = "^2.1.0"
djangorestframework = "^3.11.0"
djangorestframework-camel-case = "^1.2.0"
drf-spectacular = "^0.22.0"
gunicorn = "^20.0.4"
python = ">=3.10,<3.11"
ramos = "^1.4.0"
requests = "^2.23.0"
whitenoise = "^6.0.0"

[tool.poetry.dev-dependencies]
bandit = "^1.6.2"
black = "^22.1"
flake8 = "^4.0.1"
ipython = "^8.0.0"
isort = "^5.2.2"
model_bakery = "^1.0.2"
mypy = "^0.950"
pdbpp = "^0.10.2"
pytest = "^7.0.0"
pytest-cov = "^3.0.0"
pytest-deadfixtures = "^2.1.0"
pytest-django = "^4.1.0"
pytest-env = "^0.6.2"
pytest-pythonpath = "^0.7.3"
pytest-vcr = "^1.0.2"
safety = "^1.8.5"
vcrpy = "^4.1.0"

[tool.black]
line-length = 79
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )
)
'''

[tool.isort]
line_length = 79
known_first_party = "tempracatao"
multi_line_output = 3
use_parentheses = true
include_trailing_comma = true
skip_glob = [
    "**migrations**"
]

[tool.coverage.run]
omit = [
    'src/manage.py',
    '**asgi.py',
    '**wsgi.py',
    '**migrations**',
    '**tests**',
    '**settings**',
]
branch = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
